requirements.txt
flask>=3.1
flask-sqlalchemy>=3.1
flask-wtf>=1.2
flask-migrate>=4.0
qrcode>=8.0
pillow>=10.0
gunicorn>=21.2
psycopg2-binary

# dev / test
pytest>=8
pytest-flask>=1.3
pytest-cov>=5
ruff>=0.5
factory-boy>=3.3
